자동 로또 만드는 메서드 수정 자동 로또를 만드는 기능을 User에서 LottoMachine으로 이전
